add: fastify-cli dev script
